separate elimination reason enum
