+++ vali_objects/enums/elimination_reason_enum.py
@@ -1,0 +1,28 @@
+from enum import Enum
+
+
+# ==================== Elimination Types ====================
+
+class EliminationReason(Enum):
+    """Reasons for miner elimination."""
+    ZOMBIE = "ZOMBIE"
+    INACTIVE = "INACTIVE"
+    PLAGIARISM = "PLAGIARISM"
+    MAX_TOTAL_DRAWDOWN = "MAX_TOTAL_DRAWDOWN"
+    FAILED_CHALLENGE_PERIOD_TIME = "FAILED_CHALLENGE_PERIOD_TIME"
+    FAILED_PROBATION_TIME = "FAILED_PROBATION_TIME"
+    FAILED_CHALLENGE_PERIOD_DRAWDOWN = "FAILED_CHALLENGE_PERIOD_DRAWDOWN"
+    FAILED_CHALLENGE_PERIOD_INTRADAY_DRAWDOWN = "FAILED_CHALLENGE_PERIOD_INTRADAY_DRAWDOWN"
+    FAILED_CHALLENGE_PERIOD_EOD_DRAWDOWN = "FAILED_CHALLENGE_PERIOD_EOD_DRAWDOWN"
+    FAILED_FUNDED_PERIOD_INTRADAY_DRAWDOWN = "FAILED_FUNDED_PERIOD_INTRADAY_DRAWDOWN"
+    FAILED_FUNDED_PERIOD_EOD_DRAWDOWN = "FAILED_FUNDED_PERIOD_EOD_DRAWDOWN"
+    LIQUIDATED = "LIQUIDATED"
+    DEREGISTERED = "DEREGISTERED"
+
+    @property
+    def is_intraday_drawdown(self):
+        return self in (self.FAILED_CHALLENGE_PERIOD_INTRADAY_DRAWDOWN, self.FAILED_FUNDED_PERIOD_INTRADAY_DRAWDOWN)
+
+    @property
+    def is_eod_drawdown(self):
+        return self in (self.FAILED_CHALLENGE_PERIOD_EOD_DRAWDOWN, self.FAILED_FUNDED_PERIOD_EOD_DRAWDOWN)
